extract wait_for_all into futures/wait_for_all.hpp.
