+++ include/boost/thread/future.hpp
@@ -23,6 +23,7 @@
 #include <boost/thread/exceptional_ptr.hpp>
 #include <boost/thread/futures/future_error_code.hpp>
 #include <boost/thread/futures/is_future_type.hpp>
+#include <boost/thread/futures/wait_for_all.hpp>
 #include <boost/thread/lock_algorithms.hpp>
 #include <boost/thread/lock_types.hpp>
 #include <boost/thread/mutex.hpp>
@@ -1082,60 +1083,6 @@ namespace boost
     struct is_future_type<shared_future<T> > : true_type
     {
     };
-
-    template<typename Iterator>
-    typename boost::disable_if<is_future_type<Iterator>,void>::type wait_for_all(Iterator begin,Iterator end)
-    {
-        for(Iterator current=begin;current!=end;++current)
-        {
-            current->wait();
-        }
-    }
-
-#ifdef BOOST_NO_CXX11_VARIADIC_TEMPLATES
-    template<typename F1,typename F2>
-    typename boost::enable_if<is_future_type<F1>,void>::type wait_for_all(F1& f1,F2& f2)
-    {
-        f1.wait();
-        f2.wait();
-    }
-
-    template<typename F1,typename F2,typename F3>
-    void wait_for_all(F1& f1,F2& f2,F3& f3)
-    {
-        f1.wait();
-        f2.wait();
-        f3.wait();
-    }
-
-    template<typename F1,typename F2,typename F3,typename F4>
-    void wait_for_all(F1& f1,F2& f2,F3& f3,F4& f4)
-    {
-        f1.wait();
-        f2.wait();
-        f3.wait();
-        f4.wait();
-    }
-
-    template<typename F1,typename F2,typename F3,typename F4,typename F5>
-    void wait_for_all(F1& f1,F2& f2,F3& f3,F4& f4,F5& f5)
-    {
-        f1.wait();
-        f2.wait();
-        f3.wait();
-        f4.wait();
-        f5.wait();
-    }
-#else
-    template<typename F1, typename... Fs>
-    void wait_for_all(F1& f1, Fs&... fs)
-    {
-        bool dummy[] = { (f1.wait(), true), (fs.wait(), true)... };
-
-        // prevent unused parameter warning
-        (void) dummy;
-    }
-#endif // !defined(BOOST_NO_CXX11_VARIADIC_TEMPLATES)
 
     template<typename Iterator>
     typename boost::disable_if<is_future_type<Iterator>,Iterator>::type wait_for_any(Iterator begin,Iterator end)
